Updated to "No Queen in the MobilityArea". https://github.com/official-stockfish/Stockfish/commit/45072612d4ddc1483581eb7799d270778e71d5e3

+++ src/tbcore.c
@@ -1481,7 +1481,7 @@ static uint8_t *decompress_pairs(struct PairsData *d, size_t idx)
       litidx -= d->sizetable[block++] + 1;
   }
 
-  uint32_t *ptr = (uint32_t *)(d->data + (block << d->blocksize));
+  uint32_t *ptr = (uint32_t *)(d->data + ((uint64_t)block << d->blocksize));
 
   int m = d->min_len;
   uint16_t *offset = d->offset;
